gtk: Fix .solid-csd styling again
In the previous commit ce5adf6066ee08bf3fd7945563ca70728c8cd466, we were
unable to resize the window at the top edge of some GTK apps.

This commit adjusts the styling to be able to resize the window at the
top edge.

Also, this changes the resizing border width from 4px to 2px, as well as
Metacity theme and Xfwm theme.

https://github.com/nana-4/materia-theme/issues/429