gtk: Change background/surface rule
Based on Material Design guidelines, this tries not to change the
background colors when widgets are at the same elevation. For example,
sidebar-specific and view-specific background colors no longer exist;
they will have different background colors depending on the elevation.

To make it easier to achieve, this does the following:

- Treat frames as cards and add background-color to them.
- Remove background-color from elevation-less widgets as much as
  possible.